[azdo] reorder publish artifacts step
Context: https://jopepper.visualstudio.com/Jon%20Peppers%20OSS/_build/results?buildId=292

I was publishin artifacts immediately after build. This was so I didn't have to wait so long to grab the files -- the tests take ~10 min on Windows.

It turns out this is a bad idea if a test fails and you have to re-run. Any reruns will completely fail with:

    Artifact windows already exists for build 292.

So I moved the test after the test runs, which is probably the correct way anyway.