Fix `boots` command-line arguments
I noticed this on CI:

    > dotnet Boots/bin/$(Configuration)/netcoreapp3.1/Boots.dll --preview Xamarin.Android
    At least one of --url, --stable, or --preview must be used

What? `--preview` was passed???

This broke in 674d09b6e, when I used the wrong API for validating
incoming command-line options...

I will delist 1.0.3.551.

To prevent this in the future, I added `MainTests` that actually check
that the `static Task Main()` returns the proper output and works
correctly.